requirements.txt

Flask==1.0.2
Flask-Excel==0.0.7
pyexcel-handsontable==0.0.2
pyexcel-sortable==0.0.2
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
